requirements.txt
albumentations
dataclasses
einops
future
h5py
http://pai-vision-data-hz.cn-hangzhou.oss-cdn.aliyun-inc.com/third_party/nvidia_dali_cuda100-0.25.0-1535750-py3-none-manylinux2014_x86_64.whl
#http://pai-vision-data-hz.oss-cn-zhangjiakou.aliyuncs.com/third_party/mmcv_full-1.4.4%2Btorch1.8.0%2Bcu102-cp36-cp36m-manylinux1_x86_64.whl
# http://pai-vision-data-hz.oss-cn-zhangjiakou.aliyuncs.com/third_party/mmcv_full-1.3.18%2Btorch1.5.0%2Bcu101-cp36-cp36m-manylinux1_x86_64.whl
#http://pai-vision-data-hz.oss-cn-zhangjiakou.aliyuncs.com/third_party/mmcv_full-1.2.0%2Btorch1.7.0%2Bcu101-cp36-cp36m-manylinux1_x86_64.whl
json_tricks
numpy
opencv-python-headless==4.4.0.40
oss2
packaging
Pillow<=6.2.2
pycocotools
pytorch_metric_learning==0.9.89
scikit-image
sklearn
tensorboard
thop
timm==0.4.9
xtcocotools
yacs
